create Component
